Fix build: Resolve data module type mismatches and missing DTOs (Round 3)

+++ android/data/src/main/java/com/malsync/android/data/local/entity/AnimeEntity.kt
@@ -31,6 +31,6 @@ data class AnimeEntity(
     val studios: List<String>,
     val provider: SyncProvider,
     val providerIds: Map<SyncProvider, String>,
-    val nextEpisodeAiring: Int?,
+    val nextEpisodeAiring: Long?,
     val lastUpdated: Long
 )
